fixup! [ADD] stock_price_br and CST-level creditable taxes
