+++ l10n_br_fiscal/views/cst_view.xml
@@ -46,6 +46,7 @@
                       <field name="name" />
                       <field name="cst_type" />
                       <field name="tax_group_id" />
+                      <field name="default_creditable_tax" />
                     </group>
                 </sheet>
             </form>
