+++ l10n_br_fiscal/views/tax_view.xml
@@ -35,6 +35,7 @@
                 <field name="tax_group_id" />
                 <field name="cst_in_id" />
                 <field name="cst_out_id" />
+                <field name="creditable_tax" />
             </tree>
         </field>
     </record>
