+++ l10n_br_stock_account/models/__init__.py
@@ -1,12 +1,9 @@
 # Copyright (C) 2015  Renato Lima - Akretion
 # License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
 
-from . import cst
-from . import tax
 from . import res_company
 from . import stock_rule
 from . import stock_picking_type
 from . import stock_picking
-from . import stock_price_mixin
 from . import stock_move
 from . import procurement_group
