--- l10n_br_stock_account/views/tax_view.xml
@@ -1,14 +1,0 @@
-<odoo>
-
-    <record id="tax_form_inherit" model="ir.ui.view">
-        <field name="name">l10n_br_fiscal.tax.form.inherit</field>
-        <field name="inherit_id" ref="l10n_br_fiscal.tax_form" />
-        <field name="model">l10n_br_fiscal.tax</field>
-        <field name="arch" type="xml">
-            <field name="cst_out_id" position="after">
-                <field name="creditable_tax" />
-            </field>
-        </field>
-    </record>
-
-</odoo>
